✨ Add `geometry` dtype, PostGIS support, and pages navigation. (#203)
- **Add the dtype `geometry` (and `geography`).**
  The new data type `geometry` adds support for syncing GIS data (e.g. WKB/WKT, GeoJSON, `shapely` objects, and `GeoDataFrames`).

  ```python
  import meerschaum as mrsm
  pipe = mrsm.Pipe(
      'demo', 'geometry',
      instance='sql:local',
      columns={
          'primary': 'id',
      },
      dtypes={
          'geom': 'geometry',
          'id': 'int',
      },
  )

  ```

  The `geometry` dtype syntax supports constraints for Geometry type and/or CRS (SRID):

  ```yaml
  dtypes:
    id: int
    geom: geometry[MultiLineString, 4326]
  ```

  Syncing a `GeoDataFrame` (without specifying an explicit dtype) will detect any CRS and geometry type:

  ```python
  import meerschaum as mrsm
  import geopandas as gpd

  path = 'path/to/shapefile.shp'
  gdf = gpd.read_file(path)

  pipe = mrsm.Pipe('demo', 'shapefile', instance='sql:local')
  success, msg = pipe.sync(gdf)

  print(pipe.dtypes['geometry'])
  # geometry[Point, 6570]
  ```

- **Add the SQLConnector flavor `postgis`.**
  The new flavor `postgis` (built atop the `postgresql` flavor) allows pipes to take natively support `GEOMETRY` (and `GEOGRAPHY`) types.

- **Add a pages sidebar to the Web Console.**
  Clicking the Meerschaum logo on the Web Console will show the pages navigation sidebar, allowing you to easily expand the web app. Custom pages added via `@web_page` are grouped by plugin. You can override this behavior with the `page_group` parameter:

  ```python
  from meerschaum.plugins import web_page, dash_plugin

  @dash_plugin
  def init_dash(dash_app):
      import dash.html as html

      @web_page('users', page_group='Settings', login_required=True)
      def my_new_settings_page():
          return dbc.Container([
              html.P("A new page `/dash/users/` added to the Settings page group.")
          ])
  ```

- **Add the property `instance_keys` to `api` connectors.**
  The optional property `instance_keys` determines the value of `instance_keys` to be sent alongside pipe requests.

- **Insert the Web Console navbar into custom pages.**
  Custom pages added via `@web_page` will now include the simple navbar by default, to more tightly integrate custom pages into the Web Console. This behavior may be disabled by setting `skip_navbar=True` in `@web_page`:

  ```python
  # example.py

  from meerschaum.plugins import web_page, dash_plugin

  @dash_plugin
  def init_app(dash_app):
      import dash_bootstrap_components as dbc
      import dash.html as html

      @web_page('example-page', skip_navbar=True)
      def example_page():
          return dbc.Container(html.H1('Hello, world!'))
  ```

- **Create `INT` columns for dtypes `int32`, `SMALLINT` for `int16`.**
  The `SQLConnector` now maps the Pandas dtypes `int32` to `INT` and `int16` (and `int8`) to `SMALLINT` rather than defaulting to `BIGINT` for everything.

- **Fix serialization of `valkey` pipes without indices.**
  Pipes synced without `columns` now correctly serialize documents' keys.

- **Add API endpoints for clearing pipes and chunk bounds.**
  The endpoints `/pipes/{connector_keys}/{metric_key}/{location_key}/clear` and `/pipes/{connector_keys}/{metric_key}/{location_key}/chunk_bounds` now allow API users to clear pipes (rather than using the legacy actions endpoint) and get the values from `pipe.get_chunk_bounds()`.

- **Skip venv locking on Windows.**
- **Shrink `full` Docker image size.**

+++ tests/test_pipes_dtypes.py
@@ -977,6 +977,8 @@ def test_enforce_false(flavor: str):
     """
     Test `enforce=False` behavior.
     """ 
+    if flavor not in ('api', 'mssql', 'timescaledb', 'sqlite'):
+        return
     conn = conns[flavor]
     pipe = mrsm.Pipe('test', 'enforce', instance=conn)
     pipe.delete()
@@ -984,12 +986,13 @@ def test_enforce_false(flavor: str):
         'test', 'enforce',
         instance=conn,
         enforce=False,
+        dtypes={'dt': 'datetime64[ns]'},
         columns={
             'datetime': 'dt',
         },
     )
     docs = [
-        {'dt': datetime(2024, 12, 26, tzinfo=timezone.utc), 'num': 1},
+        {'dt': datetime(2024, 12, 26), 'num': 1},
     ]
     success, msg = pipe.sync(docs, debug=debug)
     assert success, msg
@@ -998,7 +1001,7 @@ def test_enforce_false(flavor: str):
     assert df['num'][0] == 1
     
     new_docs = [
-        {'dt': datetime(2024, 12, 26, tzinfo=timezone.utc), 'num': 2},
+        {'dt': datetime(2024, 12, 26), 'num': 2},
     ]
     success, msg = pipe.sync(new_docs, debug=debug)
     assert success, msg
@@ -1066,3 +1069,129 @@ def test_explicit_precision_scale_numeric_sql(flavor):
     success, msg = pipe.sync(docs, debug=debug)
     df = pipe.get_data()
     assert str(df['val'][0]) == '123.46'
+
+@pytest.mark.parametrize("flavor", get_flavors())
+def test_geometry_dtype(flavor: str):
+    """
+    Test syncing the `geometry` dtype.
+    """
+    conn = conns[flavor]
+    shapely = mrsm.attempt_import('shapely')
+    pipe = mrsm.Pipe(
+        'test', 'geometry',
+        instance=conn,
+    )
+    pipe.delete()
+    pipe = mrsm.Pipe(
+        'test', 'geometry',
+        instance=conn,
+        columns={'primary': 'id'},
+        dtypes={'id': 'int', 'geom': 'geometry'},
+    )
+
+    geom = shapely.MultiLineString([[[0, 0], [1, 2]], [[4, 4], [5, 6]]])
+    geom_str = '01050000000200000001020000000200000000000000000000000000000000000000000000000000F03F00000000000000400102000000020000000000000000001040000000000000104000000000000014400000000000001840'
+    docs = [
+        {'id': 1, 'geom': geom},
+    ]
+    new_docs = [
+        {'id': 1, 'geom': geom_str},
+    ]
+
+    success, msg = pipe.sync(docs, debug=debug)
+    assert success, msg
+
+    df = pipe.get_data()
+    assert df['geom'][0] == geom
+
+    success, msg = pipe.sync(df, debug=debug)
+    assert success, msg
+    assert pipe.get_rowcount() == len(docs)
+
+    success, msg = pipe.sync(new_docs, debug=debug)
+    assert success, msg
+    assert pipe.get_rowcount() == len(docs)
+
+
+@pytest.mark.parametrize("flavor", get_flavors())
+def test_geography_dtype(flavor: str):
+    """
+    Test syncing the `geography` dtype.
+    """
+    conn = conns[flavor]
+    shapely = mrsm.attempt_import('shapely')
+    pipe = mrsm.Pipe(
+        'test', 'geography',
+        instance=conn,
+    )
+    pipe.delete()
+    pipe = mrsm.Pipe(
+        'test', 'geography',
+        instance=conn,
+        columns={'primary': 'id'},
+        dtypes={'id': 'int', 'geog': 'geography[SRID=4326]'},
+    )
+
+    geog = shapely.MultiLineString([[[0, 0], [1, 2]], [[4, 4], [5, 6]]])
+    geog_str = '01050000000200000001020000000200000000000000000000000000000000000000000000000000F03F00000000000000400102000000020000000000000000001040000000000000104000000000000014400000000000001840'
+    docs = [
+        {'id': 1, 'geog': geog},
+    ]
+    success, msg = pipe.sync(docs, debug=debug)
+    assert success, msg
+
+    df = pipe.get_data()
+    assert df['geog'][0] == geog
+
+    success, msg = pipe.sync(df, debug=debug)
+    assert success, msg
+    assert pipe.get_rowcount() == len(docs)
+
+    new_docs = [
+        {'id': 1, 'geog': geog_str},
+    ]
+    success, msg = pipe.sync(new_docs, debug=debug)
+    assert success, msg
+    assert pipe.get_rowcount() == len(docs)
+
+
+@pytest.mark.parametrize("flavor", get_flavors())
+def test_geometry_custom_srid(flavor: str):
+    """
+    Test syncing the `geometry` dtype with a custom SRID.
+    """
+    conn = conns[flavor]
+    shapely = mrsm.attempt_import('shapely')
+    pipe = mrsm.Pipe(
+        'test', 'geometry', 'srid',
+        instance=conn,
+    )
+    pipe.delete()
+    pipe = mrsm.Pipe(
+        'test', 'geometry', 'srid',
+        instance=conn,
+        columns={'primary': 'id'},
+        dtypes={'id': 'int', 'geom': 'geometry[MultiLineString,2001]'},
+    )
+
+    geom = shapely.MultiLineString([[[0, 0], [1, 2]], [[4, 4], [5, 6]]])
+    geom_str = '01050000000200000001020000000200000000000000000000000000000000000000000000000000F03F00000000000000400102000000020000000000000000001040000000000000104000000000000014400000000000001840'
+    docs = [
+        {'id': 1, 'geom': geom},
+    ]
+    success, msg = pipe.sync(docs, debug=debug)
+    assert success, msg
+
+    df = pipe.get_data()
+    assert df['geom'][0] == geom
+
+    success, msg = pipe.sync(df, debug=debug)
+    assert success, msg
+    assert pipe.get_rowcount() == len(docs)
+
+    new_docs = [
+        {'id': 1, 'geom': geom_str},
+    ]
+    success, msg = pipe.sync(new_docs, debug=debug)
+    assert success, msg
+    assert pipe.get_rowcount() == len(docs)
